Update android-studio-emulator.md
I did try to use the hint in line 23, and the script in lines 26 and 27 but it was not working on my side as I'm on Ubuntu and there is no .bash_profile but .profile.
I'm still struggling to make the adb command working. I'm assuming lines 32 and 33 are incorrect and need to be tested. Sorry I tried my best, I'm new with Linux!